% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndTransform.R
\name{TestAndTransform}
\alias{TestAndTransform}
\title{Test and Transform Numeric Columns in a Dataframe}
\usage{
TestAndTransform(dataframe, min_non_na_prop = 0.5, exclude = c())
}
\arguments{
\item{dataframe}{A dataframe to process.}

\item{min_non_na_prop}{The minimum proportion of non-NA values required in a column for it to be considered for testing and transformation.
Default is 0.5, meaning at least 50\% of the values in the column must be non-NA.}

\item{exclude}{A vector of column names to exclude from testing and transformation.}
}
\value{
A dataframe with numeric columns transformed as necessary, excluding specified columns.
}
\description{
Iterates over each numeric column in a dataframe, tests each for normality using the Shapiro-Wilk test,
applies the best normalization transformation if needed, based on a user-specified minimum proportion of non-NA values,
and allows excluding specific columns by name. Returns a new dataframe with transformed columns.
}
\examples{
data(iris)
# Transform iris dataset with the default minimum non-NA proportion, excluding the 'Species' column
transformed_iris_excluded <- TestAndTransform(iris, exclude = c("Species"))
}
